Add paraneter to get_base_table function
The get_base_table function is to take an additional optional parameter 
for the database link. Where this is set, the link is carried over from 
the calling function to perform the various MySQL queries. Where it is 
not set then the function will call admin_db_connect (as per the 
previous default behaviour).