Various changes
1. Add optional function parameters for updating charsets/collation and 
for optimisation. Enclose the appropriate sections of code in conditions 
relating to these parameters.
2. Add functionality to handle a non-empty value of the $dbid parameter 
(was probably removed on a previous change). Where $dbid is set, the 
connection information is obtained from the global $dbinfo array, rather 
than by calling the default admin_db_connect and admin_db_name 
functions.
3. Add a new user message at the top to indicate which database is being 
processed (helpful when running a command to process all databases).
4. Move the generation of the user message indicating the table being 
processed to the top of the table processing loop.